telegram-bot/requirements.txt
python-telegram-bot>=20.0
yfinance>=0.2.0
pandas>=2.0.0
numpy>=1.24.0
